Move location tests to new file
